Ajout d'utilisateurs avec privilèges (admin/employé), attribution de tâches, intégration d'images dans la base de données

+++ controller/ControllerVoiture.php
@@ -9,35 +9,75 @@ class ControllerVoiture extends controller {
     public function index(){
        
         $voiture = new Voiture;
-       /*  $selectCategorie = $voiture->voitureCategorie(); */
-        $select = $voiture->select();
+        $selectCategorie = $voiture->selectCategorie();
+      /*   $select = $voiture->select(); */
         
        /*  $categorie = new Categorie;
         $selectCategorie = $categorie->selectId($select['categorie_id']);
      */
-        return Twig::render('voiture-index.php', ['voitures'=>$select]);
+        return Twig::render('voiture-index.php', ['voitures'=>$selectCategorie]);
         
        
 
     }
 
      public function create(){
-
+        CheckSession::sessionAuth();
+        if($_SESSION['privilege'] == 1 || $_SESSION['privilege'] == 2){
         $categorie = new Categorie;
         $select = $categorie->select();
         return Twig::render('voiture-create.php', ['categories'=>$select]);
-        
+        }
+        RequirePage::url('login'); 
     } 
 
     public function store(){
-        print_r($_POST);
+       /*  print_r($_POST);
                 
         $voiture = new Voiture;
         $insert = $voiture->insert($_POST);
      
         print_r($selectId); 
 
-         RequirePage::url('voiture/show/'.$insert);  
+         RequirePage::url('voiture/show/'.$insert);   */
+         $voiture = new Voiture;
+
+         if (isset($_FILES['photo']) && $_FILES['photo']['error'] === UPLOAD_ERR_OK) {
+            $target_dir = "/PHP/tp1-PHP-session3/uploads/img/";
+            $target_file = $_SERVER['DOCUMENT_ROOT'].$target_dir .$_FILES['photo']["name"];
+            
+            $check = getimagesize($_FILES['photo']["tmp_name"]);
+            
+            if ($check !== false) {
+               
+                move_uploaded_file($_FILES['photo']["tmp_name"], $target_file);
+        
+               
+                $_POST['photo'] = $_FILES['photo']["name"];
+                           
+             
+            } else {
+                echo "Le fichier n'est pas une image valide.";
+                die();
+               
+            }
+
+                      
+        }
+       /*  var_dump([$_POST]); */
+        unset($_POST['submit']);
+        $insert = $voiture->insert($_POST);
+        RequirePage::url('voiture/show/'.$insert); 
+
+
+       /*  $_POST['photo'] = $_FILES['photo']['name']; */
+           
+/*             $voiture = new Voiture; */
+
+
+           
+         
+
 
        
 
@@ -72,10 +112,14 @@ class ControllerVoiture extends controller {
     }
 
     Public function destroy(){
-        
+        CheckSession::sessionAuth();
+        if($_SESSION['privilege'] == 2){
         $voiture = new Voiture;
         $update = $voiture->delete($_POST['id']);
         RequirePage::url('voiture/index');
+        }else{
+            RequirePage::url('login');
+        }
     }
